arreglado error de locacion de las clases de selenium
