+++ src/main/java/junit/lvazquez/atc02_BusquedaDirectaProductoExistente.java
@@ -1,0 +1,4 @@
+package junit.lvazquez;
+
+public class atc02_BusquedaDirectaProductoExistente {
+}
