+++ src/test/java/selenium/lvazquez/act01TestSuite.java
@@ -1,4 +1,4 @@
-package selenium;
+package selenium.lvazquez;
 
 import org.junit.runners.Suite;
 import org.junit.runner.RunWith;
